Remove dependency on ghoneycutt-common
We had some problems with the ghoneycutt-common module and prefer to avoid using
"glue" modules wherever possible.  As ghoneycutt-common is only used for its
`mkdir_p` function, we don't see it as a big issue to remove the dependency,
even if it means that directories will have to be created from somewhere else.